requirements.txt
djangorestframework
django-celery-results
drf-spectacular
djangorestframework-simplejwt
drf-spectacular